+++ arm/ui/utils.py
@@ -20,6 +20,7 @@ from werkzeug.routing import ValidationError
 from flask.logging import default_handler  # noqa: F401
 
 import arm.config.config as cfg
+from arm.config import config_utils
 from arm.ui import app, db
 from arm.models import models
 from arm.ui.metadata import tmdb_search, get_tmdb_poster, tmdb_find, call_omdb_api
@@ -645,7 +646,7 @@ def build_arm_cfg(form_data, comments):
         if key == "csrf_token":
             continue
         # Add any grouping comments
-        arm_cfg += arm_yaml_check_groups(comments, key)
+        arm_cfg += config_utils.arm_yaml_check_groups(comments, key)
         # Check for comments for this key in comments.json, add them if they exist
         try:
             arm_cfg += "\n" + comments[str(key)] + "\n" if comments[str(key)] != "" else ""
@@ -657,54 +658,8 @@ def build_arm_cfg(form_data, comments):
             arm_cfg += f"{key}: {post_value}\n"
         except ValueError:
             # Test if value is Boolean
-            arm_cfg += arm_yaml_test_bool(key, value)
+            arm_cfg += config_utils.arm_yaml_test_bool(key, value)
     app.logger.debug("save_settings: FINISH")
-    return arm_cfg
-
-
-def arm_yaml_test_bool(key, value):
-    """
-    we need to test if the key is a bool, as we need to lower() it for yaml\n\n
-    or check if key is the webserver ip. \nIf not we need to wrap the value with quotes\n
-    :param key: the current key
-    :param value: the current value
-    :return: the new updated arm.yaml config with new key: values
-    """
-    if value.lower() == 'false' or value.lower() == "true":
-        arm_cfg = f"{key}: {value.lower()}\n"
-    else:
-        # If we got here, the only key that doesn't need quotes is the webserver key
-        # everything else needs "" around the value
-        if key == "WEBSERVER_IP":
-            arm_cfg = f"{key}: {value.lower()}\n"
-        else:
-            # This isn't intended to be safe, it's to stop breakages - replace all non escaped quotes with escaped
-            escaped = re.sub(r"(?<!\\)[\"\'`]", r'\"', value)
-            arm_cfg = f"{key}: \"{escaped}\"\n"
-    return arm_cfg
-
-
-def arm_yaml_check_groups(comments, key):
-    """
-    Check the current key to be added to arm.yaml and insert the group
-    separator comment, if the key matches\n
-    :param comments: comments dict, containing all comments from the arm.yaml
-    :param key: the current post key from form.args
-    :return: arm.yaml config with any new comments added
-    """
-    comment_groups = {'COMPLETED_PATH': "\n" + comments['ARM_CFG_GROUPS']['DIR_SETUP'],
-                      'WEBSERVER_IP': "\n" + comments['ARM_CFG_GROUPS']['WEB_SERVER'],
-                      'SET_MEDIA_PERMISSIONS': "\n" + comments['ARM_CFG_GROUPS']['FILE_PERMS'],
-                      'RIPMETHOD': "\n" + comments['ARM_CFG_GROUPS']['MAKE_MKV'],
-                      'HB_PRESET_DVD': "\n" + comments['ARM_CFG_GROUPS']['HANDBRAKE'],
-                      'EMBY_REFRESH': "\n" + comments['ARM_CFG_GROUPS']['EMBY']
-                                      + "\n" + comments['ARM_CFG_GROUPS']['EMBY_ADDITIONAL'],
-                      'NOTIFY_RIP': "\n" + comments['ARM_CFG_GROUPS']['NOTIFY_PERMS'],
-                      'APPRISE': "\n" + comments['ARM_CFG_GROUPS']['APPRISE']}
-    if key in comment_groups:
-        arm_cfg = comment_groups[key]
-    else:
-        arm_cfg = ""
     return arm_cfg
